backend/requirements.txt
 fastapi==0.115.0
uvicorn==0.30.0
sqlalchemy==2.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.0
passlib[bcrypt]==1.7.4
python-jose==3.3.0
slowapi==0.1.9
pydantic[email]==2.0.0
celery==5.3.0
redis==5.0.0
